Add documentation for Windows Subsystem for Linux
Documentation on the steps needed to configure WSL in order to get to
running the local development server. Where possible, the new
documentation references existing documentation (such as how to install
WSL); however, some sections needed some specific modifications for WSL.
Overall, the goal was to supplement but not replace other documentation.

Resolves: AT-5900