Add note on PostgreSQL port conflicts
Some users may already have PostgreSQL running somewhere. It seems that
the Ubuntu installer for PostgreSQL just chooses an available port (such
as 5433) when 5432 is in use. This can result in unexpected behavior.
This more generic wording doesn't preclude the use of othter PostgreSQL
instances, just that they shouldn't be bound to port 5432.