Update to Python 3.8 (#2108)
* Remove invalid argument to Formatter constructor

In Python 3.7, the first argument to logging.Formatter.__init__ is not
validated by default; however, in Python 3.8 and newer, the fmt string
is validated. Since `self` is incorrectly being passed as the value for
the `fmt` parameter, this results in a TypeError when the validation
happens since `JsonFormatter` is not a `str`. Passing `self` seems to be
a mistake.

* Update dependencies for Python 3.8

* Update secrets baseline

There are a few high-entropy strings and new phrases containing "secret"
or "password" that likely need to be ignored.

* Update pipeline to use Python 3.8

* Bump to Python 3.8.7

* Reword WSL install instructions

* Fix ops Docker container python symlink

The previous symlink created a link from a non-existent target to a
misleading name. It seems that perhaps the intent was to make it so
invoking `python` invoked the `python3.7` binary; however:
  - Python 3.7.3's `make install` places data in /usr/local
  - The arguments are backwards so python3.7 is a broken link
  - Neither the base RHEL install or `make install` create a file in
    /usr/bin nor /usr/local/bin with the name `python`

Flipping the arguments, using `/usr/local/` and now referencing
`python3.8` makes a `python` executable available on the PATH; however,
it is unclear whether this is actually needed by any scripts since it
would not have worked before.

* Bump pip version to 21.*

In 2018, pip switched to using date-based versioning rather than
semantic versioning, so the version jumped from 10 to 18. `pip` 20.3
switched to the new dependency resolver; this does not seem to cause any
issues with our `setup` script or dependency resolution.

This jump is therefore not quite as signficant as it may seem and
keeping pip up-to-date with the Python release is a worthwhile
endeavor.