Ensure coverage exclusions are in-sync
When files are omitted from the pytest-cov report and SonarCloud has not
been configured to exclude those files from coverage analysis, it can
result in incorrect reports on test coverage as well as in PR test
failures (in situations where the files modified are the ones excluded
only from pytest-cov reports). Keeping the two lists in-sync is
necessary to ensure the reports are accurate.

Ticket: AT-5996

Co-authored-by: Jason Burkert (Hunter Strategy) <77642966+jasonburkert-ccpo@users.noreply.github.com>