Regenerate static assets when testing
Since static/assets/* are in the .gitignore file, they get carried over
from branch-to-branch as untracked files. In general this is fine since
they should not be checked in; however, in the scenario where the `yarn
build` fails and the assets cannot be generated, the tests will continue
without issue since the files were untracked and carried over. Then when
pushing to GitHub, the CI job will run and the tests will fail with
obscure error messages about being unable to find files. This is due to
the fact generating them is broken but that couldn't be detected
locally. Forcing a full regeneration with each test run should provide a
clear error message if that is broken.

Ticket: AT-5966