Add missing argument to deployer script
There is a @click.option for an argument that is not being passed to the
deployer function. While the argument does not seem to be used, I am
unsure if it is being passed/provided in any way, so this was the
quickest fix to resolve it. Long-term, it should be compared to the git
sha argument to determine which we need.